third_party/WebKit/LayoutTests/inspector/elements/styles-1/css-live-edit.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/elements-test.js"></script>
<script src="../../../http/tests/inspector/live-edit-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<link rel="stylesheet" href="../styles/resources/css-live-edit.css">
<div id=foo></div>
<script>

function test()
{
    TestRunner.runTestSuite([
        function testLiveEdit(next)
        {
            SourcesTestRunner.showScriptSource("css-live-edit.css", didShowResource);

            function didShowResource(sourceFrame)
            {
                TestRunner.addSniffer(SDK.CSSModel.prototype, "_fireStyleSheetChanged", didEditResource);
                SourcesTestRunner.replaceInSource(sourceFrame, "font-size: 12px;", "font-size: 20px;");
            }

            function didEditResource()
            {
                ElementsTestRunner.selectNodeAndWaitForStylesWithComputed("foo", didSelectElement);
            }

            function didSelectElement()
            {
                ElementsTestRunner.dumpSelectedElementStyles(false, true);
                next();
            }
        }
    ]);
};

</script>

</head>

<body onload="runTest()">
<p>Tests that styles are updated when live-editing css resource.</p>

</body>
</html>
