third_party/WebKit/LayoutTests/inspector/elements/styles-1/css-live-edit-expected.txt

Tests that styles are updated when live-editing css resource.


Running: testLiveEdit
display: block;
font-size: 20px;
[expanded] 
element.style  { ()
